nabot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
